Added snowflake settings screen. Snowflake variables moved to used integer values coz of normalized screen.

+++ app/src/main/java/dev/jatzuk/snowwallpaper/opengl/SnowfallRenderer.kt
@@ -63,8 +63,8 @@ class SnowfallRenderer(private val context: Context) : GLSurfaceView.Renderer {
 
         ratio = if (width > height) width.toFloat() / height else height.toFloat() / width
 
-        OpenGLWallpaperService.width = width.toFloat()
-        OpenGLWallpaperService.height = height.toFloat()
+        OpenGLWallpaperService.width = width
+        OpenGLWallpaperService.height = height
 
         isSnowfallBackgroundProgramUsed = preferenceRepository.getIsSnowfallEnabled()
         isSnowflakeProgramUsed = preferenceRepository.getIsSnowflakeEnabled()
